added static_cast

+++ app/jni/src/core/graphics/VertexAttributes.cpp
@@ -4,20 +4,20 @@
 
 #include "VertexAttributes.h"
 #ifndef  __ANDROID__
-#include <GL/glew.h>
+#include "GL/glew.h"
 #else
 #include <GLES3/gl31.h>
 #endif
 void core::rend::VertexAttributes::add(i32 size,i32 stride,i32 offset)
 {
-    glVertexAttribPointer(attributes, size, GL_FLOAT, GL_FALSE, stride * sizeof(float), (void*)(offset * sizeof(float)));
+    glVertexAttribPointer(attributes, size, GL_FLOAT, GL_FALSE,static_cast<int>(stride * sizeof(float)), (void*)(offset * sizeof(float)));
     glEnableVertexAttribArray(attributes);
     attributes++;
 }
 
 void core::rend::VertexAttributes::add(i32 size, size_t typeSize, i32 offset)
 {
-    glVertexAttribPointer(attributes, size, GL_FLOAT, GL_FALSE,  typeSize, (void*)(offset));
+    glVertexAttribPointer(attributes, size, GL_FLOAT, GL_FALSE,  static_cast<int>(typeSize), (void*)(offset));
     glEnableVertexAttribArray(attributes);
     attributes++;
 }
